backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-dotenv==1.0.1
scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.2
joblib==1.4.2
xgboost==2.1.2
